Fix compiler warning in Py CFC.

+++ compiler/python/src/cfc/_cfc.c
@@ -177,7 +177,7 @@ static PyObject*
 S_CFCHierarchy_get_dest(PyObject *wrapper, PyObject *unused) {
     CHY_UNUSED_VAR(unused);
     const char *dest = CFCHierarchy_get_dest(S_to_Hierarchy(wrapper));
-    return PyUnicode_DecodeASCII(dest, strlen(dest), NULL);
+    return PyUnicode_DecodeASCII(dest, (Py_ssize_t)strlen(dest), NULL);
 }
 
 static PyObject*
